fix: keep team summaries within typed endpoint
