--- apps/api/src/handlers/time-entries/summary/team/get.ts
@@ -1,123 +1,0 @@
-import { Result } from "better-result";
-import { and, eq, gte, lte, sql } from "drizzle-orm";
-import { t } from "elysia";
-
-import { member, user } from "@/api/db/auth-schema";
-import { timeEntries, workspaceMembers } from "@/api/db/schema";
-import { createSafeHandler } from "@/api/lib/api-handlers";
-import { addDays, parseIsoDateLocal } from "@/api/lib/dates";
-import { HandlerError } from "@/api/lib/errors/tagged-errors";
-import { LIMITS } from "@/api/lib/limits";
-
-const MAX_SUMMARY_DAYS = 31;
-const MAX_TEAM_SUMMARY_ROWS =
-  LIMITS.workspaceMembersCount * MAX_SUMMARY_DAYS;
-
-const teamTimeEntrySummaryQuerySchema = t.Object({
-  dateFrom: t.String({ format: "date" }),
-  dateTo: t.String({ format: "date" }),
-});
-
-const readTeamTimeEntrySummary = createSafeHandler(
-  {
-    description:
-      "Summarize each matter member's time by day for a bounded date range.",
-    permissions: { timeEntry: ["approve"] },
-    mcp: { type: "capability", reason: "billing_admin" },
-    access: "read",
-    query: teamTimeEntrySummaryQuerySchema,
-  },
-  async function* ({ query, safeDb, session, user: currentUser, workspaceId }) {
-    const from = parseIsoDateLocal(query.dateFrom);
-    const to = parseIsoDateLocal(query.dateTo);
-
-    if (!from || !to || to < from || to > addDays(from, MAX_SUMMARY_DAYS - 1)) {
-      return Result.err(
-        new HandlerError({
-          status: 400,
-          message: `Summary range must contain between 1 and ${MAX_SUMMARY_DAYS} days`,
-        }),
-      );
-    }
-
-    const rows = yield* Result.await(
-      safeDb((tx) =>
-        tx
-          .select({
-            userId: workspaceMembers.userId,
-            name: user.name,
-            email: user.email,
-            image: user.image,
-            dateWorked: timeEntries.dateWorked,
-            totalMinutes: sql<number>`coalesce(sum(${timeEntries.durationMinutes}), 0)::int`,
-          })
-          .from(workspaceMembers)
-          .innerJoin(
-            member,
-            and(
-              eq(member.userId, workspaceMembers.userId),
-              eq(member.organizationId, session.activeOrganizationId),
-            ),
-          )
-          .innerJoin(user, eq(user.id, member.userId))
-          .leftJoin(
-            timeEntries,
-            and(
-              eq(timeEntries.workspaceId, workspaceId),
-              eq(timeEntries.userId, workspaceMembers.userId),
-              gte(timeEntries.dateWorked, query.dateFrom),
-              lte(timeEntries.dateWorked, query.dateTo),
-            ),
-          )
-          .where(eq(workspaceMembers.workspaceId, workspaceId))
-          .groupBy(
-            workspaceMembers.userId,
-            user.name,
-            user.email,
-            user.image,
-            timeEntries.dateWorked,
-          )
-          .orderBy(user.name, workspaceMembers.userId, timeEntries.dateWorked)
-          .limit(MAX_TEAM_SUMMARY_ROWS),
-      ),
-    );
-
-    const membersById = new Map<
-      string,
-      {
-        userId: string;
-        name: string;
-        email: string;
-        image: string | null;
-        daily: { dateWorked: string; totalMinutes: number }[];
-      }
-    >();
-    let viewerTotalMinutes = 0;
-    for (const row of rows) {
-      const teamMember = membersById.get(row.userId) ?? {
-        userId: row.userId,
-        name: row.name,
-        email: row.email,
-        image: row.image,
-        daily: [],
-      };
-      if (row.dateWorked !== null) {
-        teamMember.daily.push({
-          dateWorked: row.dateWorked,
-          totalMinutes: row.totalMinutes,
-        });
-        if (row.userId === currentUser.id) {
-          viewerTotalMinutes += row.totalMinutes;
-        }
-      }
-      membersById.set(row.userId, teamMember);
-    }
-
-    return Result.ok({
-      viewerTotalMinutes,
-      members: [...membersById.values()],
-    });
-  },
-);
-
-export default readTeamTimeEntrySummary;
